update account + get cinema + get showtime

+++ src/main/java/com/example/moviebookingbackend/repository/TicketRepository.java
@@ -5,6 +5,6 @@ import com.example.moviebookingbackend.model.Ticket;
 
 import java.util.List;
 
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
-//    List<Ticket> findByAccountId(Long accountId);
+public interface TicketRepository extends JpaRepository<Ticket, Integer> {
+//    List<Ticket> findByAccountId(Integer accountId);
 }
